Time: 267 ms (5.03%) | Memory: 19.2 MB (5.95%) - LeetSync

+++ 907-koko-eating-bananas/koko-eating-bananas.py
@@ -11,7 +11,7 @@ class Solution:
             
             total = 0
             for p in piles:
-                total += math.ceil(float(p)/k)
+                total += math.ceil(int(p)/k)
 
             if total > h:
                 l = k + 1
